requirements.txt
newsapi-python
openai
awscli
sqlalchemy
sqlalchemy-utils
apache-airflow-providers-common-sql
apache-airflow-providers-postgres